app/requirements.txt
Django==3.1.5
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2-binary==2.8.5
